% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotModelPrediction}
\alias{plotModelPrediction}
\title{Plot the predictions by the model
One plot per measured node
Plot with error bars if available, and give absolute value}
\usage{
plotModelPrediction(model, targets, readouts = "all", plotsPerFrame = 4,
  log_axis = F)
}
\arguments{
\item{model}{An MRAmodel object}

\item{targets}{A matrix describing the conditions to be simulated}

\item{readouts}{Vector with the name of the nodes that should be used as readouts}

\item{plotsPerFrame}{}

\item{log_axis}{Boolean, whether the ordinate axis should be in log scale}
}
\value{
Nothing
}
\description{
Plot the predictions by the model
One plot per measured node
Plot with error bars if available, and give absolute value
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}

