% Generated by roxygen2 (4.0.1): do not edit by hand
\name{printParameters}
\alias{printParameters}
\title{Print the value of each path from the model, and add the profile likelihood infos if they are provided}
\usage{
printParameters(model_description)
}
\arguments{
\item{model_description}{An MRAmodel object}
}
\value{
Nothing
}
\description{
Print the value of each path from the model, and add the profile likelihood infos if they are provided
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}

