% Generated by roxygen2 (4.0.1): do not edit by hand
\name{selectMinimalModel}
\alias{selectMinimalModel}
\title{Selection of a minimal model by the removal of non significant links with a Chi^2 test}
\usage{
selectMinimalModel(model_description, accuracy = 0.95)
}
\arguments{
\item{model_description}{A list describing the model, as the one produced by createModel or importModel}

\item{accuracy}{Probability of the confidence interval for the Chi^2 test. The link can be removed if 0 is included in this interval.}
}
\value{
An MRAmodel object of the reduced model with the data
}
\description{
Selection of a minimal model by the removal of non significant links with a Chi^2 test
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}

